+++ languages/elixir/exercises/concept/anonymous-functions/.docs/hints.md
@@ -1,0 +1,42 @@
+### General
+
+- Make use of [anonymous functions][anon-fns].
+- Use a closure to reference the variable from the outer scope.
+
+### 1. Create an adder
+
+- Your function should accept one integer parameter, then use that parameter as a reference in the returned anonymous function.
+- The anonymous function should also accept an integer parameter.
+- The two parameters should be added in the anonymous function scope.
+
+### 2. Create a subtractor
+
+- Your function should accept one integer parameter, then use that parameter as a reference in the returned anonymous function.
+- The anonymous function should also accept an integer parameter.
+- The two parameters should be subtracted in the anonymous function scope.
+
+### 3. Create a multiplier
+
+- Your function should accept one integer parameter, then use that parameter as a reference in the returned anonymous function.
+- The anonymous function should also accept an integer parameter.
+- The two parameters should be multiplied in the anonymous function scope.
+
+### 4. Create a divider
+
+- Your function should accept one integer parameter, then use that parameter as a reference in the returned anonymous function.
+- The anonymous function should also accept an integer parameter.
+- The two parameters should be divided, using integer division, in the anonymous function scope.
+
+### 5. Create a list appender
+
+- Your function should accept one integer parameter, then use that parameter as a reference in the returned anonymous function.
+- The anonymous function should also accept an integer parameter.
+- The two parameters should be list concatenated in the anonymous function scope.
+
+### 6. Create a string appender
+
+- Your function should accept one integer parameter, then use that parameter as a reference in the returned anonymous function.
+- The anonymous function should also accept an integer parameter.
+- The two parameters should be string concatenated in the anonymous function scope.
+
+[anon-fns]: https://elixir-lang.org/getting-started/basic-types.html#anonymous-functions
