+++ languages/elixir/exercises/concept/anonymous-functions/.meta/config.json
@@ -1,0 +1,9 @@
+{
+  "authors": [
+    {
+      "github_username": "neenjaw",
+      "exercism_username": "neenjaw"
+    }
+  ],
+  "language_versions": ">=1.10"
+}
