+++ languages/elixir/exercises/concept/anonymous-functions/.meta/design.md
@@ -1,0 +1,37 @@
+# Design
+
+## Goal
+
+The goal of this exercise is to teach students how to use anonymous functions, return functions from functions, use closures in code.
+
+## Learning objectives
+
+After completing this exercise, the student should:
+
+- Know how to write [anonymous functions][fns].
+- Know how to return a function from a function.
+- Know how to use closures in elixir.
+
+## Out of scope
+
+- pattern matching
+- function overloading
+
+## Concepts
+
+The concepts this exercise unlocks are:
+
+- `anonymous-functions`: know of the existence of anonymous functions and know how to write them.
+- `closures`: know how to use closures in Elixir
+
+## Representer
+
+This exercise does not require any specific representation logic to be added to the [representer][representer].
+
+## Analyzer
+
+This exercise does not require any specific logic to be added to the [analyzer][analyzer].
+
+[analyzer]: https://github.com/exercism/elixir-analyzer
+[representer]: https://github.com/exercism/elixir-representer
+[fns]: https://elixir-lang.org/getting-started/basic-types.html#anonymous-functions
