+++ languages/elixir/exercises/concept/anonymous-functions/lib/closure_maker.ex
@@ -1,25 +1,25 @@
 defmodule ClosureMaker do
   def make_adder(number) do
-    fn x -> x + number end
+    raise "Please implement the make_adder/1 function."
   end
 
   def make_subtractor(number) do
-    fn x -> x - number end
-  end
-
-  def make_divider(number) do
-    fn x -> div(x, number) end
+    raise "Please implement the make_subtractor/1 function."
   end
 
   def make_multiplier(number) do
-    fn a -> a * number end
+    raise "Please implement the make_multiplier/1 function."
+  end
+
+  def make_divider(number) do
+    raise "Please implement the make_divider/1 function."
   end
 
   def make_list_appender(list) do
-    fn l -> l ++ list end
+    raise "Please implement the make_list_appender/1 function."
   end
 
   def make_string_appender(string) do
-    fn s -> s <> string end
+    raise "Please implement the make_string_appender/1 function."
   end
 end
