+++ languages/elixir/exercises/concept/anonymous-functions/mix.exs
@@ -5,7 +5,7 @@ defmodule ClosureMaker.MixProject do
     [
       app: :anonymous_functions,
       version: "0.1.0",
-      elixir: "~> 1.10",
+      # elixir: "~> 1.10",
       start_permanent: Mix.env() == :prod,
       deps: deps()
     ]
