+++ languages/elixir/exercises/concept/conditionals/.meta/design.md
@@ -2,7 +2,7 @@
 
 ## Goal
 
-The goal of this exercise to teach students how to use conditional control flow structures
+The goal of this exercise is to teach students how to use conditional control flow structures
 
 ## Learning objectives
