+++ languages/elixir/reference/README.md
@@ -257,11 +257,13 @@ TODO: Flesh out how the above general concepts apply to concept exercises in the
 
 The concept exercises use the following concepts:
 
-| concept           | interpretation                                                                                 |
-| ----------------- | ---------------------------------------------------------------------------------------------- |
-| `basics`          | Introduction to functions, modules, variables, returning values, integers, invoking functions. |
-| `booleans`        | Introduction to the boolean type and strict boolean operators -- and/2, or/2, not/1            |
-| `conditionals`    | Intro to elixir `cond/1` function.                                                             |
-| `atoms`           | Intro to elixir atom type.                                                                     |
-| `lists`           | Introduction to the lists type basic list functions -- hd/1, tl/1, length/1, in/2              |
-| `string-literals` | Introduction to strings in code enclosed by double quotes.                                     |
+| concept               | interpretation                                                                                 |
+| --------------------- | ---------------------------------------------------------------------------------------------- |
+| `basics`              | Introduction to functions, modules, variables, returning values, integers, invoking functions. |
+| `booleans`            | Introduction to the boolean type and strict boolean operators -- and/2, or/2, not/1            |
+| `conditionals`        | Intro to elixir `cond/1` function.                                                             |
+| `atoms`               | Intro to elixir atom type.                                                                     |
+| `lists`               | Introduction to the lists type basic list functions -- hd/1, tl/1, length/1, in/2              |
+| `string-literals`     | Introduction to strings in code enclosed by double quotes.                                     |
+| `anonymous-functions` | Intro to anonymous functions, functions as data                                                |
+| `closures`            | How to implement closures in elixir                                                            |
