+++ languages/elixir/reference/concepts.csv
@@ -3,3 +3,4 @@ basics,
 booleans,
 conditionals,
 lists,
+anonymous-functions,
